update: clean up overall marketing page structure and shared components
