+++ apps/website/src/routes/privacy.tsx
@@ -1,14 +1,16 @@
-import { createFileRoute } from '@tanstack/react-router';
-import { LegalMdxPage, MarketingLayout } from '@/components/marketing';
-import PrivacyContent from '@/content/legal/privacy.mdx';
+import { createFileRoute } from "@tanstack/react-router";
+import { LegalMdxLayout } from "@/components/marketing/layouts/legal-mdx-layout";
+import { MarketingLayout } from "@/components/marketing/layouts/marketing-layout";
+import PrivacyContent from "@/content/legal/privacy.mdx";
 
-export const Route = createFileRoute('/privacy')({
+export const Route = createFileRoute("/privacy")({
   head: () => ({
     meta: [
-      { title: 'Privacy Policy | Fabricator' },
+      { title: "Privacy Policy | Fabricator" },
       {
-        name: 'description',
-        content: 'Privacy policy for fabricator.site and the Fabricator website.',
+        name: "description",
+        content:
+          "Privacy policy for fabricator.site and the Fabricator website.",
       },
     ],
   }),
@@ -18,13 +20,13 @@ export const Route = createFileRoute('/privacy')({
 function Privacy() {
   return (
     <MarketingLayout active="privacy" legal>
-      <LegalMdxPage
+      <LegalMdxLayout
         eyebrow="Privacy"
         title="Privacy Policy"
         intro="This privacy policy explains which personal data may be processed when you visit fabricator.site."
       >
         <PrivacyContent />
-      </LegalMdxPage>
+      </LegalMdxLayout>
     </MarketingLayout>
   );
 }
