+++ apps/website/src/styles/app.css
@@ -163,6 +163,35 @@ html > body[data-scroll-locked] {
   text-underline-offset: 4px;
 }
 
+.fabricator-site-bg {
+  background:
+    radial-gradient(circle at 8% 18%, rgb(34 211 238 / 22%), transparent 30%),
+    radial-gradient(circle at 84% 0%, color-mix(in oklab, var(--primary) 34%, transparent), transparent 30%),
+    radial-gradient(circle at 70% 48%, rgb(15 23 42 / 28%), transparent 38%),
+    linear-gradient(180deg, #080808 0%, #101010 48%, #151515 100%);
+}
+
+.fabricator-site-bg::after {
+  position: absolute;
+  inset: 0;
+  background:
+    linear-gradient(90deg, rgb(255 255 255 / 2%) 1px, transparent 1px),
+    linear-gradient(180deg, rgb(255 255 255 / 1.5%) 1px, transparent 1px);
+  background-size: 72px 72px;
+  content: '';
+  mask-image: radial-gradient(circle at 50% 18%, black, transparent 68%);
+  opacity: 0.18;
+}
+
+@media (max-width: 640px) {
+  .fabricator-site-bg {
+    background:
+      radial-gradient(circle at 0% 12%, rgb(34 211 238 / 20%), transparent 36%),
+      radial-gradient(circle at 100% 0%, color-mix(in oklab, var(--primary) 30%, transparent), transparent 36%),
+      linear-gradient(180deg, #080808 0%, #101010 52%, #151515 100%);
+  }
+}
+
 .legal-mdx > :first-child {
   display: none;
 }
